Safari-609.1.11.1/LayoutTests/media/session/controls-existence.html

<!DOCTYPE html>
<html>
<head>
    <script src=../video-test.js></script>
    <script type="text/javascript">
        var session;
    
        function doTest () 
        {
            session = new MediaSession("content");
            run("session = new MediaSession('content')");
            testExpected("session.controls", null, '!=');

            session = new MediaSession("transient");
            run("session = new MediaSession('transient')");
            testExpected("session.controls", null);

            session = new MediaSession("transient-solo");
            run("session = new MediaSession('transient-solo')");
            testExpected("session.controls", null);

            session = new MediaSession("ambient");
            run("session = new MediaSession('ambient')");
            testExpected("session.controls", null);

            endTest();
        }
    </script>
</head>
<body onload="doTest()">
    <p>Only Content Media Sessions should have media remote controls after creation.</p>
</body>
</html>
