Safari-609.1.11.1/LayoutTests/media/session/controls-existence-expected.txt

Only Content Media Sessions should have media remote controls after creation.

RUN(session = new MediaSession('content'))
EXPECTED (session.controls != 'null') OK
RUN(session = new MediaSession('transient'))
EXPECTED (session.controls == 'null') OK
RUN(session = new MediaSession('transient-solo'))
EXPECTED (session.controls == 'null') OK
RUN(session = new MediaSession('ambient'))
EXPECTED (session.controls == 'null') OK
END OF TEST
